Refactored filters.py into a directory filters containing filter and utils python files
